+++ tensorboard/components/experimental/plugin_lib/BUILD
@@ -4,21 +4,11 @@ load("//tensorboard/defs:defs.bzl", "tf_ts_library")
 
 licenses(["notice"])  # Apache 2.0
 
-# Generating module friendly message.ts.
-# The genrule creates a new version of message.ts that strips `namespace {` and
-# `} // namespace`.
-# TODO(tensorboard-team): tf_ts_library and tf_web_library do not interoperate
-# yet we do have a requirement to make the plugin_lib use tf_ts_library. Remove
-# this when the two can interop or tf_web_library gets deprecated.
-genrule(
-    name = "message_lib",
+tf_ts_library(
+    name = "message",
     srcs = [
-        "//tensorboard/components/experimental/plugin_util:message.ts",
+        "message.ts",
     ],
-    outs = ["message.ts"],
-    cmd = """
-        cat $< | sed -e "s|^namespace .*||" | sed -e "s|^} // namespace .*||" >$@
-    """,
 )
 
 # TODO(psybuzz): create a NPM package when a better requirement comes up using
@@ -29,6 +19,8 @@ tf_ts_library(
         "index.ts",
         "plugin-guest.ts",
         "runs.ts",
-        ":message_lib",
+    ],
+    deps = [
+        ":message",
     ],
 )
